refactored gameMode, added Survival
